workflow: refactor provider and extract html template
This refactoring separates UI concerns from the provider logic and improves code maintainability by following the established pattern in the extension.

1. Provider Refactoring:
   - Rename SidebarWorkflowWebviewProvider.ts to WorkflowWebviewProvider.ts.
   - Update the provider to use external HTML generation.

2. HTML Generation:
   - Create WorkflowHtmlGenerator.ts to handle variable injection.
   - Extract HTML, CSS, and JS into resources/webview/workflow-tree-template.html.

3. Constants Integration:
   - Centralize workflow command IDs and UI messages in constants.ts.
   - Protocol message IDs in the webview are kept as string literals for consistency with other webview providers.

All changes are verified with a successful build.
